build: work in progress

+++ packages/nx-gradle/src/generators/kotlin-multiplatform/files/android-app/build.gradle.kts
@@ -35,6 +35,6 @@ android {
         targetCompatibility = JavaVersion.toVersion(javaVersion)
     }
     kotlin {
-        jvmToolchain(javaVersion)
+        jvmToolchain(javaVersion.toInt())
     }
 }
